[optim][BE] split test file into logical parts: SWA, LR, optim
[ghstack-poisoned]
